% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrase_to_pkg.R
\name{phrase_to_pkgs}
\alias{phrase_to_pkgs}
\title{phrase_to_pkg}
\usage{
phrase_to_pkgs(phrase, screen_dump = TRUE, exact = TRUE, n = 10,
  open_url = FALSE)
}
\arguments{
\item{phrase}{of one or more words.}

\item{screen_dump}{If \code{TRUE}, package titles and descriptions are output to
screen in a nicely formatted manner.}

\item{exact}{If \code{TRUE}, only packages with titles or descriptions which
exactly match the given phrase are returned; otherwise the \code{n} best
matches are returned.}

\item{n}{For \code{exact = FALSE}, the number of best matches to be returned.}

\item{open_url}{If \code{TRUE}, open CRAN web pages of matching packages.}
}
\value{
A \code{tibble} containing package names, titles, and descriptions.
}
\description{
Finds the package that most closely matches a given text phrase.
}
\note{
This function is *not* intended to be used like \link{phrase_to_pkg},
rather it is just a helper function that dumps results to screen.
}
