% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrase_to_pkg.R
\name{textsearch}
\alias{textsearch}
\title{textsearch}
\usage{
textsearch(phrase)
}
\arguments{
\item{phrase}{of one or more words.}
}
\value{
At present nothing; just prints package title and description to
screen.
}
\description{
Finds the package that most closely matches a given text phrase.
}
\note{
This function is intended to extract a single package from which to
start *flipping*.
}
